[Feature] 다른사람이 merge할때 컨트롤러의 스웨거 코드 날라간거 복구, 커뮤니티 게시글, 댓글 추상화
